Fix: Change userId type from Long to Integer to match database SERIAL type

+++ src/main/java/murach/business/User.java
@@ -10,7 +10,7 @@ public class User implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "userid")
-    private Long userId;
+    private Integer userId;  // Changed from Long to Integer to match SERIAL type
     
     @Column(name = "firstname", nullable = false, length = 50)
     private String firstName;
@@ -33,11 +33,11 @@ public class User implements Serializable {
         this.email = email;
     }
     
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
     
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
